requirements.txt
psycopg2==2.7.7
django==2.1.3
django-allauth
django-ckeditor
whoosh
django-haystack
